feat: enhance cart and order management with admin interfaces, serializers, and views

+++ apps/orders/models.py
@@ -27,5 +27,5 @@ class Order(models.Model):
 class OrderItem(models.Model):
     order = models.ForeignKey(Order, related_name='order_items', on_delete=models.CASCADE)
     product = models.ForeignKey(Product, related_name='order_items', on_delete=models.CASCADE)
-    quantity = models.IntegerField()
+    quantity = models.PositiveBigIntegerField()
     price = models.FloatField()
